refactor: moved axios
